Fix unable to unblock XMPP buddy
If the user blocked an XMPP buddy that was not in the buddy list,
Pidgin sent the XID with the resource to the server's block list.

If the user tried to unblock that buddy after adding them to the
buddy list, Pidgin sent the XID without the resource to the server
for unblocking.

This resulted in users unable to unblock their buddies.

This patch fixes this situation by sending just the normalized XID
(i.e. without the resource) to the server for blocking and
unblocking.

Testing Done:
Tested on a Prosody XMPP server by adding a buddy and unblocking them.

Bugs closed: PIDGIN-16414

Reviewed at https://reviews.imfreedom.org/r/1479/