Don't link with libgadu unnecessarily
The default if-found action of `AC_CHECK_LIB` is to add the library to `LIBS`,
but that means it's used for _all_ linking. So a GPL-compliant libgadu gets
linked with libpurple, finch, pidgin, and all plugins instead of just the
protocol plugin.

Testing Done:
Compiled and checked that there was no `libgadu.so` in `ldd libpurple.so` and a few other built things, but it was still there for the Gadu-Gadu protocol plugin.

Reviewed at https://reviews.imfreedom.org/r/2684/